requirements.txt
falcon>=0.1.9
pastedeploy>=1.3.3
simplejson>=2.0.9
pbr>=0.6,!=0.7,<1.0
six>=1.7.0
stevedore>=0.14
gunicorn>=19.1.0
oslo.config>=1.2.1
ujson>=1.33
Pyparsing>=2.0.3
influxdb>=0.1.12
